requirements.txt
Pillow
pydub
tqdm
